base: added national roaming for Orange in Austria
http://jira.omnirom.org/browse/OMNI-376

Users of Orange in Autria (MCC 232 MNC 05) which is now a subcompany
of 3AT can roam on 3AT (MNC10) and also on t-mobile (MNC03) for 2G

Change-Id: I593c1bdeea3ef712541b4962369a02b2d7c86c5b

National roaming info for Drei in Austria

3AT roams in T-Mobile AT networks

Change-Id: I915edccf92bcbb088fa25b958eb345c624f61597

core: added national roaming info for Bob in Austria.

Users of Bob in Austria(MCC 232 MNC 11)
  can roam on A1 Austria (MNC 01)

Change-Id: I12dfa3d63643af1026dcfea28a4c9d7acec5f9e9

National roaming info for yesss! in Austria

Change-Id: I930198d3cb311b23752a8c2f54852ebd48d65b86